Move PlatformType to API and expose it for extensions
