+++ api/src/main/java/org/geysermc/geyser/api/util/PlatformType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2022 GeyserMC. http://geysermc.org
+ * Copyright (c) 2019-2023 GeyserMC. http://geysermc.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,13 +23,11 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-package org.geysermc.common;
+package org.geysermc.geyser.api.util;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
+/**
+ * Represents the platform Geyser is running on.
+ */
 public enum PlatformType {
     ANDROID("Android"),
     BUNGEECORD("BungeeCord"),
@@ -40,4 +38,12 @@ public enum PlatformType {
     VELOCITY("Velocity");
 
     private final String platformName;
+
+    PlatformType(String platformName) {
+        this.platformName = platformName;
+    }
+
+    public String getPlatformName() {
+        return platformName;
+    }
 }
