+++ core/src/main/java/org/geysermc/connector/GeyserConnector.java
@@ -26,7 +26,7 @@
 package org.geysermc.connector;
 
 import org.geysermc.api.Geyser;
-import org.geysermc.common.PlatformType;
+import org.geysermc.geyser.api.util.PlatformType;
 import org.geysermc.connector.network.session.GeyserSession;
 import org.geysermc.geyser.GeyserImpl;
