+++ core/src/main/java/org/geysermc/geyser/command/defaults/HelpCommand.java
@@ -25,7 +25,7 @@
 
 package org.geysermc.geyser.command.defaults;
 
-import org.geysermc.common.PlatformType;
+import org.geysermc.geyser.api.util.PlatformType;
 import org.geysermc.geyser.GeyserImpl;
 import org.geysermc.geyser.api.command.Command;
 import org.geysermc.geyser.command.GeyserCommand;
