+++ core/src/main/java/org/geysermc/geyser/dump/BootstrapDumpInfo.java
@@ -27,7 +27,7 @@ package org.geysermc.geyser.dump;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.geysermc.common.PlatformType;
+import org.geysermc.geyser.api.util.PlatformType;
 import org.geysermc.geyser.GeyserImpl;
 import org.geysermc.geyser.text.AsteriskSerializer;
